refactor(genui): isolate provider process lifecycle
